extract Index Handler
